requirements.txt
asgiref==3.8.1
Django==4.2.13
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
